requirements.txt
streamlit
rembg
opencv-python-headless
numpy
pillow